[feat] Refactor VectorDB class hierarchy for flexibility
